requirements.txt
yt-dlp
openai
python-dotenv
requests
